[DESK-163] Create a PullRequestStep model class (#352)
## [DESK-163] Create a PullRequestStep model class
