+++ src/www/modals/ProcessPR/StepChangesDescription/index.js
@@ -1,9 +1,9 @@
-const Component = require('@interface/Component');
+const StepModel = require('../StepModel');
 const Button = require('@www/components/DocForm/FormField/Button');
 const Spinner = require('@www/components/Spinner');
 const FileChange = require('@www/components/FileChange');
 
-class StepChangesDescription extends Component {
+class StepChangesDescription extends StepModel {
     get SOURCE_PATH() {
         return require.resolve('./StepChangesDescription.html');
     }
@@ -14,7 +14,6 @@ class StepChangesDescription extends Component {
         const { isLoading, fileChanges } = Object(settings);
 
         this.isLoading = new Spinner();
-        this.setCurrent(true);
         this.setButton.save(true);
 
         this.fileChanges = fileChanges || [];
@@ -22,26 +21,6 @@ class StepChangesDescription extends Component {
         this.types = {
             FileChange
         };
-    }
-
-    setCurrent(state) {
-        try {
-            if (state) {
-                this.isCurrentClass = 'current-step';
-            } else {
-                this.isCurrentClass = '';
-            }
-        } catch (err) {
-            throw new Error.Log(err);
-        }
-    }
-
-    removeError() {
-        try {
-            delete this.error;
-        } catch (err) {
-            throw new Error.Log(err);
-        }
     }
 
     get setError() {        
@@ -63,19 +42,6 @@ class StepChangesDescription extends Component {
                 }
             }
         };
-    }
-
-    resolve() {
-        try {
-            this.resolved = true;
-            this.setCurrent(false);
-            this.removeError();
-
-            // Hiding the buttons
-            Object.keys(this.setButton).map(key => this.setButton[key](false));
-        } catch (err) {
-            throw new Error.Log(err);
-        }
     }
 }
