+++ src/www/modals/ProcessPR/StepCommit/index.js
@@ -1,10 +1,10 @@
-const Component = require('@interface/Component');
+const StepModel = require('../StepModel');
 const Spinner = require('@www/components/Spinner');
 const Button = require('@www/components/DocForm/FormField/Button');
 const CommitFileChange = require('./CommitFileChange');
 const { TextArea, Input } = require('@www/components/DocForm/FormField/fields');
 
-class StepCommit extends Component {
+class StepCommit extends StepModel {
     get SOURCE_PATH() {
         return require.resolve('./StepCommit.html');
     }
@@ -24,7 +24,6 @@ class StepCommit extends Component {
             this.isLoading = new Spinner();
         }
 
-        this.setCurrent(true);
         this.setButton.nextStep(true);
         this.fileChangesMessage = `There are ${changes.length} file changes to be commited!`;
         
@@ -48,26 +47,6 @@ class StepCommit extends Component {
         this.types = {
             CommitFileChange
         };
-    }
-
-    setCurrent(state) {
-        try {
-            if (state) {
-                this.isCurrentClass = 'current-step';
-            } else {
-                this.isCurrentClass = '';
-            }
-        } catch (err) {
-            throw new Error.Log(err);
-        }
-    }
-
-    removeError() {
-        try {
-            delete this.error;
-        } catch (err) {
-            throw new Error.Log(err);
-        }
     }
 
     get setError() {        
@@ -109,19 +88,6 @@ class StepCommit extends Component {
                 }
             }
         };
-    }
-
-    resolve() {
-        try {
-            this.resolved = true;
-            this.setCurrent(false);
-            this.removeError();
-
-            // Hiding the buttons
-            Object.keys(this.setButton).map(key => this.setButton[key](false));
-        } catch (err) {
-            throw new Error.Log(err);
-        }
     }
 }
